+++ src/views/app-views/apps/documents/construction/qc-test-report/qc-test-report-list/index.js
@@ -1,19 +1,35 @@
-import React, {useState} from 'react'
-import { Card, Table, Input, Button, Menu } from 'antd';
-import ProductListData from "assets/data/product-list.data.json"
-import { EditOutlined, DeleteOutlined, SearchOutlined, PlusCircleOutlined } from '@ant-design/icons';
+import React, { useState, useEffect } from 'react'
+import { Card, Table, Button, Menu, Modal, message } from 'antd';
+import { EditOutlined, DeleteOutlined, PlusCircleOutlined } from '@ant-design/icons';
 import EllipsisDropdown from 'components/shared-components/EllipsisDropdown';
 import Flex from 'components/shared-components/Flex';
 import { useNavigate } from "react-router-dom";
+import FirestoreService from 'services/FirestoreService';
+import { doc, deleteDoc } from 'firebase/firestore';
+import { db } from 'configs/FirebaseConfig';
 import utils from 'utils'
 
 
 const QcTestList = () => {
 	const navigate = useNavigate();
-	const [list, setList] = useState(ProductListData)
+	const [list, setList] = useState()
 	const [selectedRows, setSelectedRows] = useState([])
 	const [selectedRowKeys, setSelectedRowKeys] = useState([])
 
+	useEffect(() => {
+		const fetchData = async () => {
+			try {
+				const documents = await FirestoreService.getDocuments('qcTestReport');
+				setList(documents);
+			} catch (error) {
+				console.log('Error fetching documents: ', error.message);
+			}
+		};
+		fetchData();
+	}
+	, []);
+
+	
 	const dropdownMenu = row => (
 		<Menu>
 			<Menu.Item onClick={() => viewDetails(row)}>
@@ -39,43 +55,68 @@ const QcTestList = () => {
 		navigate(`/app/apps/documents/construction/qc-test-report/qc-test-report-edit/${row.id}`)
 	}
 	
-	const deleteRow = row => {
-		const objKey = 'id'
-		let data = list
-		if(selectedRows.length > 1) {
-			selectedRows.forEach(elm => {
-				data = utils.deleteArrayRow(data, objKey, elm.id)
-				setList(data)
-				setSelectedRows([])
-			})
-		} else {
-			data = utils.deleteArrayRow(data, objKey, row.id)
-			setList(data)
-		}
-	}
+	const deleteRow = async (row) => {
+		Modal.confirm({
+		  title: 'Apakah anda yakin ingin menghapus dokumen ini ?',
+		  content: 'data yang dihapus tidak dapat dikembalikan',
+		  onOk: async () => {
+			try {
+			  // Delete the document from Firestore
+			  await deleteDoc(doc(db, 'qcTestReport', row.id));
+	  
+			  // Update the local list
+			  const objKey = 'id';
+			  let data = list;
+			  if (selectedRows.length > 1) {
+				selectedRows.forEach((elm) => {
+				  data = utils.deleteArrayRow(data, objKey, elm.id);
+				});
+				setSelectedRows([]);
+			  } else {
+				data = utils.deleteArrayRow(data, objKey, row.id);
+			  }
+			  setList(data);
+	  
+			  message.success('Document deleted successfully.');
+			} catch (error) {
+			  message.error(`Error deleting document: ${error.message}`);
+			}
+		  },
+		  onCancel: () => {
+			message.info('Deletion cancelled.');
+		  },
+		});
+	  };
 
 	const tableColumns = [
 		{
 			title: 'ID',
-			dataIndex: 'id'
+			dataIndex: 'id',
+			render: (text, record, index) => index + 1
+
 		},
 		{
 			title: 'Qc Test Report Name',
-			dataIndex: 'name',
-			sorter: (a, b) => utils.antdTableSorter(a, b, 'name')
+			dataIndex: 'qcTestReportDocsName',
+			sorter: (a, b) => utils.antdTableSorter(a, b, 'qcTestReportDocsName')
 		},
 		{
 			title: 'Project Name',
-			dataIndex: 'category',
-			sorter: (a, b) => utils.antdTableSorter(a, b, 'category')
+			dataIndex: 'projectName',
+			sorter: (a, b) => utils.antdTableSorter(a, b, 'projectName')
 		},
 		{
-			title: 'Location',
-			dataIndex: 'category',
-			sorter: (a, b) => utils.antdTableSorter(a, b, 'category')
+			title: 'File Doc',
+			dataIndex: 'documentURL',
+
+			render: (url) => (
+				<a href={url} target="_blank" rel="noopener noreferrer" download>
+				  View File
+				</a>
+			  ),
+			
+			sorter: (a, b) => utils.antdTableSorter(a, b, 'documentURL')
 		},
-		
-		
 		{
 			title: '',
 			dataIndex: 'actions',
@@ -94,20 +135,20 @@ const QcTestList = () => {
 		}
 	};
 
-	const onSearch = e => {
-		const value = e.currentTarget.value
-		const searchArray = e.currentTarget.value? list : ProductListData
-		const data = utils.wildCardSearch(searchArray, value)
-		setList(data)
-		setSelectedRowKeys([])
-	}
+	// const onSearch = e => {
+	// 	const value = e.currentTarget.value
+	// 	const searchArray = e.currentTarget.value? list : ProductListData
+	// 	const data = utils.wildCardSearch(searchArray, value)
+	// 	setList(data)
+	// 	setSelectedRowKeys([])
+	// }
 
 	return (
 		<Card>
 			<Flex alignItems="center" justifyContent="space-between" mobileFlex={false}>
 				<Flex className="mb-1" mobileFlex={false}>
 					<div className="mr-md-3 mb-3">
-						<Input placeholder="Search" prefix={<SearchOutlined />} onChange={e => onSearch(e)}/>
+						{/* <Input placeholder="Search" prefix={<SearchOutlined />} onChange={e => onSearch(e)}/> */}
 					</div>
 				</Flex>
 				<div>
